refactor(middleware): extract MCP middleware into middleware.py
Move DownstreamUnauthorizedMiddleware, ScopeFilteringMiddleware and
ToolCallLoggingMiddleware out of the growing mcp_server.py into a
dedicated gg_api_core/middleware.py (a TYPE_CHECKING import keeps the
AbstractGitGuardianFastMCP forward reference cycle-free). Pure
relocation, no behaviour change.

Refs SI-3907.
