+++ tests/test_tool_logging_middleware.py
@@ -2,7 +2,7 @@ import logging
 from types import SimpleNamespace
 
 import pytest
-from gg_api_core.mcp_server import ToolCallLoggingMiddleware
+from gg_api_core.middleware import ToolCallLoggingMiddleware
 
 
 def _ctx(name, arguments=None):
@@ -14,7 +14,7 @@ class TestToolCallLoggingMiddleware:
         async def call_next(ctx):
             return "result"
 
-        with caplog.at_level(logging.INFO, logger="gg_api_core.mcp_server"):
+        with caplog.at_level(logging.INFO, logger="gg_api_core.middleware"):
             result = await ToolCallLoggingMiddleware().on_call_tool(
                 _ctx("get_incident", {"incident_id": "123"}), call_next
             )
@@ -29,7 +29,7 @@ class TestToolCallLoggingMiddleware:
         async def call_next(ctx):
             raise ValueError("boom")
 
-        with caplog.at_level(logging.ERROR, logger="gg_api_core.mcp_server"):
+        with caplog.at_level(logging.ERROR, logger="gg_api_core.middleware"):
             with pytest.raises(ValueError, match="boom"):
                 await ToolCallLoggingMiddleware().on_call_tool(_ctx("scan_secrets"), call_next)
